Extract findnext/ package from main
Move FindNextTypist algorithm and its helpers into a dedicated package.
This is a pure algorithm with zero external dependencies, making it
the ideal first extraction in the package restructuring.

- findNextTypist() -> findnext.FindNextTypist() (exported)
- Helper functions (reverse, lookahead, contains, min, prepend) stay unexported
- Tests moved to findnext/findnext_test.go using test.Equals
